requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dj-database-url==3.0.0
Django==4.2.23
django-allauth==65.9.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.7
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.0
pytest-django==4.11.1
python-dotenv==1.1.0
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
